Move the statistics collector into the collectors package
