+++ src/main/java/sorald/cli/RepairCommand.java
@@ -17,8 +17,8 @@ import sorald.RepairStrategy;
 import sorald.SoraldConfig;
 import sorald.event.EventHelper;
 import sorald.event.EventType;
-import sorald.event.StatisticsCollector;
 import sorald.event.StatsMetadataKeys;
+import sorald.event.collectors.StatisticsCollector;
 import sorald.event.models.ExecutionInfo;
 import sorald.event.models.repair.RuleRepairStatistics;
 import sorald.sonar.RuleViolation;
