+++ src/main/java/sorald/event/collectors/StatisticsCollector.java
@@ -1,10 +1,12 @@
-package sorald.event;
+package sorald.event.collectors;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import sorald.event.SoraldEvent;
+import sorald.event.SoraldEventHandler;
 import sorald.event.models.RepairEvent;
 import sorald.event.models.miner.MinedViolationEvent;
