+++ src/test/java/sorald/processor/SoraldAbstractProcessorTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
-import sorald.event.StatisticsCollector;
+import sorald.event.collectors.StatisticsCollector;
 import spoon.Launcher;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.declaration.CtType;
